[rpc][api] Separate component related methods
	Product.check_component -> Component.filter
	Product.filter_components -> Component.filter
	Product.get_component -> Component.filter
	Product.add_component -> Component.create
	Product.update_component -> Component.update
