+++ docs/source/modules/tcms.xmlrpc.api.component.rst
@@ -1,0 +1,7 @@
+tcms\.xmlrpc\.api\.component module
+===================================
+
+.. automodule:: tcms.xmlrpc.api.component
+    :members:
+    :undoc-members:
+    :show-inheritance:
